+++ Net_Core/Web.Support/BaseControllers/AbstractCalendarController.cs
@@ -7,6 +7,7 @@ using GrowthWare.Framework.Models;
 using GrowthWare.Web.Support.Jwt;
 using GrowthWare.Web.Support.Utilities;
 using GrowthWare.Framework.Models.UI;
+using System.Threading.Tasks;
 
 namespace GrowthWare.Web.Support.BaseControllers;
 
@@ -24,9 +25,9 @@ public abstract class AbstractCalendarController : ControllerBase
     /// <returns></returns>
     [AllowAnonymous]
     [HttpGet("DeleteEvent")]
-    public ActionResult<bool> DeleteEvent(int calendarEventSeqId, string action)
+    public async Task<ActionResult<bool>> DeleteEvent(int calendarEventSeqId, string action)
     {
-        if (getEventSecurity(calendarEventSeqId))
+        if (await getEventSecurity(calendarEventSeqId))
         {
             MFunctionProfile mFunctionProfile = FunctionUtility.GetProfile(action);
             if (mFunctionProfile != null)
@@ -50,7 +51,7 @@ public abstract class AbstractCalendarController : ControllerBase
     /// <returns></returns>
     [AllowAnonymous]
     [HttpGet("GetEvent")]
-    public ActionResult<MCalendarEvent> GetEvent(string action, int id)
+    public async Task<ActionResult<MCalendarEvent>> GetEvent(string action, int id)
     {
         /** 
           * This method is not strickly necessary for the frontend.  The main
@@ -64,7 +65,7 @@ public abstract class AbstractCalendarController : ControllerBase
             MSecurityInfo mSecurityInfo = new(mFunctionProfile, mAccountProfile);
             if (mSecurityInfo.MayView)
             {
-                MCalendarEvent mRetVal = CalendarUtility.GetEvent(SecurityEntityUtility.CurrentProfile, id);
+                MCalendarEvent mRetVal = await CalendarUtility.GetEvent(SecurityEntityUtility.CurrentProfile, id);
                 HttpContext.Session.SetInt32("EditId", id);
                 return Ok(mRetVal);
             }
@@ -82,7 +83,7 @@ public abstract class AbstractCalendarController : ControllerBase
     /// <returns></returns>
     [AllowAnonymous]
     [HttpGet("GetEvents")]
-    public ActionResult<List<MCalendarEvent>> GetEvents(string action, string startDate, string endDate)
+    public async Task<ActionResult<List<MCalendarEvent>>> GetEvents(string action, string startDate, string endDate)
     {
         /** 
           * The startDate and endDate parameters are needed because the data for a calendar is
@@ -98,7 +99,7 @@ public abstract class AbstractCalendarController : ControllerBase
             {
                 DateTime mStartDate = DateTime.Parse(startDate);
                 DateTime mEndtDate = DateTime.Parse(endDate);
-                List<MCalendarEvent> mRetVal = CalendarUtility.GetEvents(SecurityEntityUtility.CurrentProfile, mFunctionProfile.Id, mStartDate, mEndtDate);
+                List<MCalendarEvent> mRetVal = await CalendarUtility.GetEvents(SecurityEntityUtility.CurrentProfile, mFunctionProfile.Id, mStartDate, mEndtDate);
                 return Ok(mRetVal);
             }
             return StatusCode(StatusCodes.Status401Unauthorized, "The requesting account does not have the correct permissions");
@@ -118,7 +119,7 @@ public abstract class AbstractCalendarController : ControllerBase
     /// </remarks>
     [AllowAnonymous]
     [HttpPost("SaveEvent")]
-    public ActionResult<MCalendarEvent> SaveEvent(UISaveEventParameters parameters)
+    public async Task<ActionResult<MCalendarEvent>> SaveEvent(UISaveEventParameters parameters)
     {
         MAccountProfile mAccountProfile = AccountUtility.CurrentProfile;
         UISaveEventParameters mParameters = parameters; // Bad practice to alter a parameter in a method.
@@ -132,10 +133,10 @@ public abstract class AbstractCalendarController : ControllerBase
             mParameters.calendarEvent.UpdatedBy = mAccountProfile.Id;
             mParameters.calendarEvent.UpdatedDate = DateTime.Now;
         }
-        if (getEventSecurity(mParameters.calendarEvent.Id, mParameters.action) && mParameters.calendarEvent.AddedBy == mAccountProfile.Id)
+        if (await getEventSecurity(mParameters.calendarEvent.Id, mParameters.action) && mParameters.calendarEvent.AddedBy == mAccountProfile.Id)
         {
             MFunctionProfile mFunctionProfile = FunctionUtility.GetProfile(mParameters.action);
-            MCalendarEvent mRetVal = CalendarUtility.SaveCalendarEvent(SecurityEntityUtility.CurrentProfile, mFunctionProfile.Id, mParameters.calendarEvent);
+            MCalendarEvent mRetVal = await CalendarUtility.SaveCalendarEvent(SecurityEntityUtility.CurrentProfile, mFunctionProfile.Id, mParameters.calendarEvent);
             return Ok(mRetVal);
         }
         return StatusCode(StatusCodes.Status401Unauthorized, "The requesting account does not have the correct permissions");
@@ -148,13 +149,13 @@ public abstract class AbstractCalendarController : ControllerBase
         return Ok(getEventSecurity(calendarEventSeqId));
     }
 
-    private bool getEventSecurity(int calendarEventSeqId, string action = null)
+    private async Task<bool> getEventSecurity(int calendarEventSeqId, string action = null)
     {
         bool mRetVal = false;
         MSecurityEntity mSecurityEntity = SecurityEntityUtility.CurrentProfile;
         if (mSecurityEntity != null)
         {
-            MCalendarEvent mCalendarEvent = CalendarUtility.GetEvent(mSecurityEntity, calendarEventSeqId);
+            MCalendarEvent mCalendarEvent = await CalendarUtility.GetEvent(mSecurityEntity, calendarEventSeqId);
             if (action == null)
             {
                 if (mCalendarEvent != null)